pyproject.toml
[tool.poetry]
name = "data_util"
version = "0.1.0"
description = ""
authors = ["Samson Fagade <sfagade@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

python-dotenv = "^1.0.1"
click = "^8.1.8"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
faker = "^33.1.0"
ruff = "^0.8.4"
[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
coverage = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

